\name{Activity10}
\docType{data}
\alias{Activity10}
\alias{DATA10}
\title{Activity patterns  of a statistician for 20 days}
\description{
  Proportion of a day in activity teaching, consulting, administrating, research,
  other wakeful activities and sleep for 20 days are given.
}
\usage{
data(Activity10)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name STATDAY.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}
\details{
 The activity of an academic statistician were divided into 
following six categories

\tabular{rll}{
  \tab    teac	 \tab teaching  \cr
  \tab    cons	 \tab consultation\cr
  \tab    admi	 \tab administration\cr
  \tab    rese	 \tab research\cr
  \tab    wake	 \tab other wakeful activities\cr
  \tab    slee	 \tab sleep
}
Data show the proportions of the 24 hours devoted to each activity,
recorded on each of 20 days, selected randomly from working days in alternate weeks,
so as to avoid any possible carry-over effects, such as short-sleep day being compensated
by make-up sleep on a succeeding day.

The six activity may be divided into two categories 'work' comprising activities
1,2,3,4: and 'leisure' comprising activities 5 and 6.

All rows sum to one.
}
\references{
Aitchison: The Statistical Analysis of Compositional Data, 1986, Data 10, pp15.
}
\note{Courtesy of J. Aitchison}
\keyword{datasets}
