\name{amounts-class}
\Rdversion{1.1}
\docType{class}
\alias{amounts-class}

\title{Class \code{"amounts"}}
\description{
Abstract class containing all amounts classes carrying total information: \code{\link{aplus}}, \code{\link{rplus}} and \code{\link{ccomp}}
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "amounts" in the signature.
}
\author{
Raimon Tolosana-Delgado
}
\seealso{
\code{\link{compositional-class}} for classes with relative information
}
\examples{
showClass("amounts")
}
\keyword{classes}
