\name{AnimalVegetation}
\docType{data}
\alias{AnimalVegetation}
\alias{Data25}
\title{Animal and vegetation measurement}
\description{Areal compositions by abundance of vegetation and animals
for 50 plots in each of regions A and B.
}
\usage{
data(AnimalVegetation)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name ANIVEG.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}

\details{
In a regional ecology study, plots of land of equal area were inspected 
and the parts of each plot which were thick or thin in vegetation and
dense or sparse in animals were identified.
From this field work the areal proportions of each plot were calculated for
the four mutually exclusive and exhaustive categories:
thick-dense, thick-sparse, thin-dense, thin-sparse.
These sets of proportions are recorded for 50 plots from each of two different
regions A and B. 

All rows sum to 1, except for some rounding errors.
}
\references{
   Aitchison: The Statistical Analysis of Compositional Data, 1986, Data 25, pp22. 
 }
\keyword{datasets}
