\name{Bayesite}
\docType{data}
\alias{Bayesite}
\alias{Data21}
\title{Permeabilities of bayesite}
\description{Relation of mechanical properties of a new fibreboard with its composition
}
\usage{
data(Bayesite)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name BAYESITE.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}

\details{In the development of bayesite, a new fibreboard,
experiments were conducted to obtain some insight into the nature of
the relationship of its permeability to the mix of its four ingredients

\tabular{rll}{
\tab A: \tab short fibres,\cr
\tab B: \tab medium fibres,\cr
\tab C: \tab long fibres,\cr
\tab D: \tab binder,
}
and the pressure of which these are bonded. The results of 21 
such experiments are reported. It is required to investigate the
dependence of permeability of mixture and bonding pressure.

All 4-part compositions sum to one.
}

\references{
   Aitchison: The Statistical Analysis of Compositional Data, 1986, Data 21, pp21.
 }
\keyword{datasets}
