\name{Blood23}
\docType{data}
\alias{Blood23}
\alias{Data23}
\title{Blood samples}
\description{
  Percentage of different leukocytes in blood samples of ten patients,
  determined by four different methods.
}
\usage{
data(Blood23)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name BLOOD.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}
\details{
  In a comparative study of four different methods of assessing leukocytes
  composition of a blood sample, aliquots of blood samples from ten patients were assessed
  by the four methods. Data show the percentage in the 40 analyses of:
  
  \tabular{rll}{
  \tab P: \tab polymorphonuclear leukocytes,\cr
  \tab S: \tab small lymphocytes,\cr
  \tab L: \tab large mononuclears.
  }

  All rows sum to 100.
}
\references{
   Aitchison: The Statistical Analysis of Compositional Data, 1986, Data 23, pp22. 
 }
\note{Courtesy of J. Aitchison}
\keyword{datasets}

