\name{Boxite}
\docType{data}
\alias{Boxite}
\alias{Data03}
\title{Compositions and depth of 25 specimens of boxite}
\description{
 A mineral compositions of 25 rock specimens of boxite type.
 Each composition consists of the percentage by weight of five minerals,
 albite, blandite, cornite, daubite, endite, as well as the depth of location.
}
\usage{
data(Boxite)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name BOXITE.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}
\details{
A mineral compositions of 25 rock specimens of boxite type are given.
 Each composition consists of the percentage by weight of five minerals,
 albite, blandite, cornite, daubite, endite, and the recorded depth of location
 of each specimen. We abbreviate the minerals names to A, B, C, D, E.
 
 All row percentage sums to 100, except the 6-th 102.3 and the 10-th 99.9.
}
\references{
   Aitchison: The Statistical Analysis of Compositional Data, 1986, Data 3, pp4.
 }
\note{Courtesy of J. Aitchison}
\keyword{datasets}
