\name{ClamEast}
\docType{data}
\alias{ClamEast}
\alias{Data14}
\title{Color-size compositions of 20 clam colonies from East Bay}
\description{
From East Bay, 20 clam colonies were randomly selected and from each a sample of clams were taken.
Each sample was sieved into three size ranges, large, medium, and small.
Then each size range was sorted by the shale colour, dark or light.
}
\usage{
data(ClamEast)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name CLAMEAST.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}
\details{
The data consist of 20 cases and \eqn{2\times 3}{2x3} variables denoted:

\tabular{rll}{
 \tab dl\tab  portion of dark large clams, \cr
 \tab dm\tab  portion of dark medium clams,\cr
 \tab ds\tab  portion of dark small clams, \cr
 \tab ll\tab  portion of light large clams, \cr
 \tab lm\tab  portion of light medium clams,\cr
 \tab ls\tab  portion of light small clams. 
}
All 6-part compositions sum to one, except for rounding errors.}
\references{
   Aitchison: The Statistical Analysis of Compositional Data, 1986, Data 14, pp18. 
 }
\note{Courtesy of J. Aitchison}
\keyword{datasets}

