\name{ClamWest}
\docType{data}
\alias{ClamWest}
\alias{Data15}
\title{Color-size compositions of 20 clam colonies from West Bay}
\description{
From West Bay, 20 clam colonies were randomly selected, and from each a sample of clams were taken.
Each sample was sieved into three size ranges, large, medium, and small.
Then each size range was sorted by the shale colour, dark or light.
}
\usage{
data(ClamWest)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name CLAMWEST.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}

\details{
The data consist of 20 cases and $2\times 3$ variables denoted:

\tabular{rll}{
 \tab dl\tab  portion of dark large clams, \cr
 \tab dm\tab  portion of dark medium clams,\cr
 \tab ds\tab  portion of dark small clams, \cr
 \tab ll\tab  portion of light large clams, \cr
 \tab lm\tab  portion of light medium clams,\cr
 \tab ls\tab  portion of light small clams.
}

All 6-part compositions sum up to one.}

\references{
   Aitchison: The Statistical Analysis of Compositional Data, 1986, Data 15, pp17. 
 }
\keyword{datasets}

