\name{compositional-class}
\Rdversion{1.1}
\docType{class}
\alias{compositional-class}

\title{Class \code{"compositional"}}
\description{
Abstract class containing all compositional classes with (at least, partly) a closed geometry: \code{\link{acomp}}, \code{\link{rcomp}} and \code{\link{ccomp}}

}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "compositional" in the signature.
}
\author{
Raimon Tolosana-Delgado
}
\seealso{
\code{\link{amounts-class}} for classes with total information
}
\examples{
showClass("compositional")
}
\keyword{classes}
