\name{Coxite}
\docType{data}
\alias{Coxite}
\alias{Data04}
\title{Compositions, depths and porosities of 25 specimens of coxite}
\description{
 A mineral compositions of 25 rock specimens of coxite type.
 Each composition consists of the percentage by weight of five minerals,
 albite, blandite, cornite, daubite, endite, the depth of location, and porosity.
}
\usage{
data(Coxite)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name BOXITE.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}

\details{
A mineral compositions of 25 rock specimens of coxite type.
 Each composition consists of the percentage by weight of five minerals,
 albite, blandite, cornite, daubite, endite,the recorded depth of location
 of each specimen, and porosity. Porosity is the percentage of void space
 that the specimen contains.
 We abbreviate the minerals names to A, B, C, D, E.
 
 All row percentage sums to 100.
}

\references{
   Aitchison: The Statistical Analysis of Compositional Data, 1986, Data 4, pp4.
 }
\keyword{datasets}

