\name{Glacial}
\docType{data}
\alias{DATA18}
\alias{Glacial}
\title{Compositions and total pebble counts of 92 glacial tills}
\description{
   In a pebble analysis of glacial tills, the total number of pebbles in each
   of 92 samples was counted and the pebbles were sorted into four categories
   red sandstone,  gray sandstone, crystalline and miscellaneous.
   The percentages of these four categories and the total pebble counts
   are recorded. 
}
\usage{
data(Glacial)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name GLACIAL.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}
\details{
   Percentages by weight in 92 samples of pebbles of glacial tills sorted 
   into four categories, red sandstone,  gray sandstone, crystalline and miscellaneous.
   The percentages of these four categories and the total pebbles counts
   are recorded. The glaciologist is interested in describing the pattern of variability
   of his data and whether the compositions are in any way related to abundance.
   All rows sum to 100, except for some rounding errors.
}
\references{
Aitchison: The Statistical Analysis of Compositional Data, 1986, Data 18, pp21.
}
\keyword{datasets}

