\name{gsiinternal1}
\alias{gsi.closespread}
\title{Internal functions of the compositions package}
\description{
  Internal functions
}
\usage{
gsi.closespread(spread)
}
\arguments{
  \item{spread}{}
}
\value{
  like spread but projected to the orthognonal complement of \code{c(1,\ldots,1)}
}
\details{

}

\note{Do not use gsi.* functions directly since they are internal
  functions of the package}

\seealso{
  \code{\link{gsi}}
}
\examples{
}
\keyword{}
