\name{gsiCoorInfo}
\alias{gsi.coorInfo}
\alias{gsi.setCoorInfo}
\alias{gsi.getCoorInfo}
\title{Internal functions of the compositions package}
\description{
  Sets or gets the triangular/barycentric coordinate system for
  the current ternary diagrams in the current plot.
}
\usage{
gsi.getCoorInfo()
gsi.setCoorInfo(...)
gsi.plots
}
\arguments{
  \item{\dots}{The info to be set typically a mean=c(1,1,1) and a scale=1}
}

\details{
  
}

\note{Do not use gsi.* functions directly since they are internal
  functions of the package}
\author{K.Gerald v.d. Boogaart \url{http://www.stat.boogaart.de}}

\seealso{
  \code{\link{gsi}}, 
}
\examples{
}
\keyword{}
