\name{gsi.eps}
\alias{gsi.eps}
\title{Internal variable: Negligible differences}
\description{
  fixes or yields the internal value for differences to be considered as zero.
}
\usage{
gsi.eps
}
\note{Do not use gsi.* functions directly since they are internal
  functions of the package}
\author{K.Gerald v.d. Boogaart \url{http://www.stat.boogaart.de}}
\seealso{
  \code{\link{gsi}}
}
\keyword{internal}
