\name{gsiinternal3}
\alias{gsi.spreadToIsoSpace}
\title{Internal functions of the compositions package}
\description{
  Internal functions
}
\usage{
gsi.spreadToIsoSpace(spread)
}
\arguments{
  \item{spread}{a matrix or a dataset of matrices}
}
\value{
  Converts a clr covariance matrix to a ilr covariance matrix
}
\details{
  
}

\note{Do not use gsi.* functions directly since they are internal
  functions of the package}

\seealso{
  \code{\link{gsi}}
}
\examples{
}
\keyword{}
