\name{Hongite}
\docType{data}
\alias{Data01}
\alias{Hongite}
\title{Compositions of 25 specimens of hongite}
\description{
 A mineral compositions of 25 rock specimens of hongite type.
 Each composition consists of the percentage by weight of five minerals,
 albite, blandite, cornite, daubite, endite.
}
\usage{
data(Hongite)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name HONGITE.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}

\details{
 A mineral compositions of 25 rock specimens of hongite type.
 Each composition consists of the percentage by weight of five minerals,
 albite, blandite, cornite, daubite, endite, which we conveniently abbreviate
 to A, B, C, D, E.  All row sums are equal to 100, except for rounding errors.
}
\references{
   Aitchison (1986): The Statistical Analysis of Compositional Data; (Data 1), pp2, 9.
 }
\keyword{datasets}
