\name{HouseholdExp}
\docType{data}
\alias{HouseholdExp}
\alias{DATA08 }
\alias{HEMF}
\title{Household Expenditures}
\description{
Household budget survey data on month expenditures of twenty men and twenty women
for four commodity groups: housing, foodstuffs, other, and services. Amounts in HK-Dollar
are given. There are 40 cases and 5 variables for 4 commodity groups and sex.
}
\usage{
data(HouseholdExp)
}
\source{ 
  Aitchison: CODA microcomputer statistical package, 1986, the file name HEMF.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}
\details{
In a sample survey of people living alone in a rented accommodation,
twenty men and twenty women were randomly selected and asked to record
over a period of one month their expenditures on the following four 
mutually exclusive and exhaustive commodity groups:
Housing, including fuel and lights, Foodstuffs, including alcohol and tobacco,
Other goods, including clothing, footwear and durable goods, and
Services, including transport and vehicles.
Amounts in HK-Dollar are given. There are 40 cases, 20 men and 20 women and
5 variables: 4 for commodity groups Housing, Food, Other, Services and the fifth sex,
$+1$ for men, $-1$ for women. Note that the data has no sum constraint. 
}
\references{
   Aitchison (1986): The Statistical Analysis of Compositional Data, (Data 08), pp13.
 }
\keyword{datasets}
