\name{Kongite}
\docType{data}
\alias{Kongite}
\alias{Data02}
\title{Compositions of 25 specimens of kongite}
\description{
 A mineral compositions of 25 rock specimens of kongite type.
 Each composition consists of the percentage by weight of five minerals,
 albite, blandite, cornite, daubite, endite.
}
\usage{
data(Kongite)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name HONGITE.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}
\details{
 A mineral compositions of 25 rock specimens of hongite type.
 Each composition consists of the percentage by weight of five minerals,
 albite, blandite, cornite, daubite, endite, which we conveniently abbreviate
 to A, B, C, D, E.  All row percentage sums to 100.
}
\references{
   Aitchison, J. (1986) The Statistical Analysis of Compositional Data, (Data 2), pp2.
 }
\keyword{datasets}
