\name{PogoJump}
\docType{data}
\alias{Data19}
\alias{PogoJump}
\title{
Honk Kong Pogo-Jumps Championship
}
\description{
   Yat, yee, sam measurements (in meters) for the final jumps of the 1985
   Honk Kong Pogo-Jumps Championship, 4 jumps of the 7 finalists.
}
\usage{
data(PogoJump)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name HKPOGO.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}
\details{
  The data consist of 28 cases: 4 jumps of the 7 finalists, and
  4 variables: Yat, Yee, Sam measurements  in meters, and  finalist -- 1 to 7.

  Pogo-Jumps is similar to the triple jump except that the competitor is
  mounted on a pogo-stik. After a pogo-up towards the starting board the total jump
  distance achieved in three consecutive bounces, known as the yat, yee and sam,
  is recorded.
}
\references{
   Aitchison, J. (1982) The Statistical Analysis of Compositional Data, (Data 19) pp21.
 }
\note{Courtesy of J. Aitchison}
\keyword{datasets}