\name{ShiftOperators}
\docType{data}
\alias{Data22}
\alias{ShiftOperators}
\title{Shifts of machine operators}
\description{Compsitions of eight-hours shifts of 27 machine operators.
}
\usage{
data(ShiftOperators)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name SHIFT.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}
\details{
A study of the activities of 27 machine operators during their
eight-hours shifts has been conducted, and proportions of time spend in the 
following categories:
\tabular{rll}{
\tab A: \tab high-quality production,\cr
\tab B: \tab low-quality production,\cr
\tab C: \tab machine setting,\cr
\tab D: \tab machine repair,
}
are recorded.
Of particular interest are any insights which such data  might give of relationships
between productive and nonproductive parts of such shifts. All compositions sum up 
one except for rounding error.
}
\references{
   Aitchison, J. (1986) The Statistical Analysis of Compositional Data, (Data 22), pp22. 
 }
\keyword{datasets}

