\name{Skulls}
\docType{data}
\alias{Data24}
\alias{Skulls}
\title{Measurement of skulls}
\description{
  Measurement in degrees of the angles N, A, B in skulls of 
     English seventeenth-century people and Naquada people.
}
\usage{
data(Skulls)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name SKULLS.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}

\details{
  As a part of a study of seventeenth-century English skulls three angles
  of a triangle in the cranium  
\tabular{rll}{
  \tab N: \tab nasial angle,\cr
  \tab A: \tab alveolar angle,\cr
  \tab B: \tab basilar angle,
  }
  were measured for 22 female and 29 male skulls.
  These, together with similar measurement of 22 female and 29 male skulls
  of the Naqada race are presented. The general objective is to investigate
  possible sex and race differences in skull shape. The angles sums in the 
  row are all equal to 180 degrees.
}
\references{
   Aitchison, J. (1986) The Statistical Analysis of Compositional Data, (Data 24), pp22. 
 }
\keyword{datasets}
