\name{SkyeAFM}
\docType{data}
\alias{Data06}
\alias{SkyeAFM}
\title{AFM compositions of 23 aphyric Skye lavas}
\description{
AFM compositions of 23 aphyric Skye lavas.
AFM diagrams formed from the relative proportions of A: alkali or Na2O + K2O,
F: Fe2O3, and M: MgO,  are common in geochemistry. 
 }
\usage{
data(SkyeAFM)
}
\source{
  Aitchison: CODA microcomputer statistical package, 1986, the file name SKYEAFM.DAT,
  here included under the GNU Public Library Licence Version 2 or newer.
}

\details{
AFM compositions of 23 aphyric Skye lavas.
AFM diagrams formed from the relative proportions of tive proportions of A: alkali or Na2O + K2O,
F: Fe2O3, and M: MgO. Adapted from Thompson, Esson and Duncan: Major element chemical variations 
in the Eocene lavas of the Isle of Skye, Scotland. All row percentage sums to 100.
}
\seealso{\code{\link[MASS]{Skye}}}
\references{
   Aitchison, J. (1986) The Statistical Analysis of Compositional Data, (Data 6) pp12.
   
   Thompson, Esson and Duncan: Major element chemical variations 
in the Eocene lavas of the Isle of Skye, Scotland. 1972, J.Petrology, 13, 219-235. 
 }
\keyword{datasets}

