\name{gsiinternal1}
\alias{gsi.closespread}
\title{Internal functions of the compositions package}
\description{
  Internal functions
}
\usage{
gsi.closespread(spread)
}
\arguments{
  \item{spread}{}
}
\value{
  like \code{spread}, but projected to the orthogonal complement of \code{c(1,\ldots,1)}
}
\note{Do not use gsi.* functions directly since they are internal
  functions of the package}
\author{K.Gerald v.d. Boogaart \url{http://www.stat.boogaart.de}}
\seealso{
  \code{\link{gsi}}
}
\keyword{internal}
