\name{gsiinternal}
\alias{gsi}
\title{Internal functions of the compositions package}
\description{
  Internal functions
}
\usage{
gsi()
}
\arguments{
  None
}
\value{
}
\details{

}

\note{Do not use gsi.* functions directly since they are internal
  functions of the package}

\seealso{
  \code{\link{gsi}}
}
\examples{
}
\keyword{}
