\name{gsiinternal3}
\alias{gsi.spreadToIsoSpace}
\title{Internal functions of the compositions package}
\description{
  Internal functions
}
\usage{
gsi.spreadToIsoSpace(spread)
}
\arguments{
  \item{spread}{a matrix or a dataset of matrices}
}
\value{
  Converts a clr covariance matrix to a ilr covariance matrix
}
\note{Do not use gsi.* functions directly since they are internal
  functions of the package}
\author{K.Gerald v.d. Boogaart \url{http://www.stat.boogaart.de}}
\seealso{
  \code{\link{gsi}}, \code{\link{clrvar2ilr}}
}
\keyword{internal}
